EE: switch to 64-bit cycle counter
Backport of upstream PCSX2 f576962dd ("EE: Switch to 64 bit cycle
counter") plus its follow-ups 14db91f75 ("mvu: fix missed 64bit
conversion") and f0723a6ec ("vu: 64bit cycle fixes"), combined into one
change since they are one logical conversion and share the savestate break.

The EE cpu cycle counter (cpuRegs.cycle) and the many counters derived from
it were u32, which wraps roughly every ~20 minutes of emulated EE time.
The wrap is the reason for the scattered "rewinding cycle" guards (the
(s32)(nextEventCycle - cycle) sign checks) and the gross +8 overflow hack in
recSkipTimeoutLoop. Widening to u64 removes that whole failure class.

Data/type changes (u32->u64, s32->s64 where the value is an absolute cycle
or a cycle delta that can now exceed 32 bits):
- cpuRegisters: cycle, sCycle[32], nextEventCycle, lastEventCycle,
  lastCOP0Cycle, lastPERFCycle[2] (eCycle[32] stays u32 - it is the event
  delta, passed as the s32 delta arg to cpuTestCycle).
- Counter/SyncCounter::startCycle, nextStartCounter, EEoCycle,
  eecount_on_last_vdec; rcntStartGate/rcntEndGate/VSyncStart/VSyncEnd take
  u64 sCycle; cpuTestCycle takes u64 startCycle and its internal compare is
  (s64) instead of (int).
- VURegs: cycle, nextBlockCycles, xgkicklastcycle; the VU pipeline sCycle
  fields (fdivPipe/efuPipe/fmacPipe/ialuPipe) and the VU interp cycle locals
  (cyclesBeforeOp, startcycles, cycle_change, vu1cycles, _cycles,
  cycle_diff, vu0_offset, runCycles, newCycle, BaseVUmicroCPU delta).
- COP0 case-9 incr -> s64 (the PERF-counter incr stays u32, matching
  upstream, since it feeds a 32-bit PERF register).
- HwRead's IntCHackCheck diff -> s64 (in the fork this lives in Hw.cpp).
- Removed the duplicate extern s32 EEsCycle / u32 EEoCycle from R3000A.h
  (declared in R5900.h), matching upstream.

JIT codegen (load/store cpuRegs.cycle and friends as 64-bit): iCOP0.cpp
recMFC0/recMTC0, microVU_Macro.inl (recV / COP2_Interlock / mVUSyncVU0),
and iR5900-32.cpp (recBranchCall, iBranchTest, recSkipTimeoutLoop,
StartRecomp) - eax/ecx/ebx -> rax/rcx/rbx and ptr32 -> ptr64 for every
cycle access. v0/GPR writebacks stay 32-bit (ptr32/edx). The recSkipTimeout
+8 overflow comment is now moot but left as-is.

Fork adaptation notes:
- The fork's recompiler file is pcsx2/x86/ix86-32/iR5900-32.cpp (upstream's
  ix86-32/iR5900.cpp); microVU/VU code uses vuRegs[] rather than VU0/VU1.
- One pre-existing fork divergence in iCOP0 recMFC0 (a shorter MFPC1 read
  sequence) was preserved.
- The "don't change order / check iR5900-X.S" struct comment is a relic;
  there is no .S file and all cycle access is by &field, so widening
  sCycle[] (which shifts later offsets) is safe.

SAVESTATE BREAK: Freeze(cpuRegs) is a raw sizeof() memory dump, so widening
these fields changes the savestate layout. g_SaveVersion bumped
0x9A56 -> 0x9A57; existing savestates (and RetroArch state slots) are
invalidated once. retro_serialize_size()'s reserved bound is unaffected
(the growth is a few dozen bytes, well within the existing headroom);
rewind/run-ahead re-serialize live so they are unaffected functionally.

Verified by inspection: every changed site traced against the three
upstream commits; struct field consistency and the core cycle arithmetic
(cpuTestCycle, counter deltas, VU sync casts) compile cleanly in isolation;
no u32* aliases &cpuRegs.cycle. NOT runtime-tested - this touches the EE/VU
dispatch hot path, so please test broadly: boot several games, run for
20+ minutes to pass the old 32-bit wrap point, exercise VU-heavy titles,
and confirm timer-sensitive games behave. Old savestates will not load
(expected).

+++ pcsx2/COP0.cpp
@@ -341,7 +341,7 @@ namespace COP0 {
 
 			case 9:
 			{
-				u32 incr = cpuRegs.cycle - cpuRegs.lastCOP0Cycle;
+				s64 incr = cpuRegs.cycle - cpuRegs.lastCOP0Cycle;
 				if (incr == 0)
 					incr++;
 				cpuRegs.CP0.n.Count += incr;
