+++ pcsx2/Hw.cpp
@@ -329,7 +329,7 @@ static __fi void IntCHackCheck(void)
 {
 	// Sanity check: To protect from accidentally "rewinding" the cyclecount
 	// on the few times nextBranchCycle can be behind our current cycle.
-	s32 diff = cpuRegs.nextEventCycle - cpuRegs.cycle;
+	s64 diff = cpuRegs.nextEventCycle - cpuRegs.cycle;
 	if (diff > 0 && (cpuRegs.cycle - cpuRegs.lastEventCycle) > 8) cpuRegs.cycle = cpuRegs.nextEventCycle;
 }
