+++ pcsx2/VUmicro.cpp
@@ -56,7 +56,7 @@ void BaseVUmicroCPU::ExecuteBlock(bool startUp)
 	}
 	else // Continue Executing
 	{
-		u32 cycle = m_Idx ? vuRegs[1].cycle : vuRegs[0].cycle;
+		u64 cycle = m_Idx ? vuRegs[1].cycle : vuRegs[0].cycle;
 		s32 delta = (s32)(u32)(cpuRegs.cycle - cycle);
 
 		if (delta > 0)
@@ -75,7 +75,7 @@ void BaseVUmicroCPU::ExecuteBlockJIT(BaseVUmicroCPU* cpu, bool interlocked)
 
 	if (stat & test) // VU is running
 	{ 
-		s32 delta = (s32)(u32)(cpuRegs.cycle - vuRegs[0].cycle);
+		s64 delta = (s64)(u64)(cpuRegs.cycle - vuRegs[0].cycle);
 		if (delta > 0)
 			cpu->Execute(CalculateMinRunCycles(delta, interlocked)); // Execute the time since the last call
 	}
