+++ pcsx2/Vif1_Dma.cpp
@@ -183,7 +183,7 @@ __fi void vif1VUFinish(void)
 	// Sync up VU1 so we don't errantly wait.
 	while (!THREAD_VU1 && (vuRegs[0].VI[REG_VPU_STAT].UL & 0x100))
 	{
-		const int cycle_diff = static_cast<int>(cpuRegs.cycle - vuRegs[1].cycle);
+		const s64 cycle_diff = static_cast<int>(cpuRegs.cycle - vuRegs[1].cycle);
 
 		if ((EmuConfig.Gamefixes.VUSyncHack && cycle_diff < vuRegs[1].nextBlockCycles) || cycle_diff <= 0)
 			break;
@@ -204,7 +204,7 @@ __fi void vif1VUFinish(void)
 
 	if (vuRegs[0].VI[REG_VPU_STAT].UL & 0x100)
 	{
-		u32 _cycles = vuRegs[1].cycle;
+		u64 _cycles = vuRegs[1].cycle;
 		vu1Finish(false);
 		if (THREAD_VU1 && !INSTANT_VU1 && (vuRegs[0].VI[REG_VPU_STAT].UL & 0x100))
 			CPU_INT(VIF_VU1_FINISH, cpuGetCycles());
